Extract legendChart fixtures from testData, also rename to legendChart
