+++ src/helpers/testData.js
@@ -309,40 +309,3 @@ export const stackedAreaData = {
     with3Sources,
     with2Sources,
 };
-
-const with6Points = () => [
-    {
-        name: 'Shiny',
-        id: 1,
-        quantity: 86,
-    },
-    {
-        name: 'Blazing',
-        id: 2,
-        quantity: 300,
-    },
-    {
-        name: 'Dazzling',
-        id: 3,
-        quantity: 276,
-    },
-    {
-        name: 'Radiant',
-        id: 4,
-        quantity: 195,
-    },
-    {
-        name: 'Sparkling',
-        id: 5,
-        quantity: 36,
-    },
-    {
-        name: 'Other',
-        id: 0,
-        quantity: 814,
-    },
-];
-
-export const legendData = {
-    with6Points,
-};
